trial

+++ perceptron/zoo/ssd_300/layers_loss_bbox/keras_layer_DecodeDetections.py
@@ -2,8 +2,8 @@ from __future__ import division
 import numpy as np
 import tensorflow as tf
 import keras.backend as K
-from keras.engine.topology import InputSpec
-from keras.engine.topology import Layer
+from tensorflow.keras.layers import InputSpec
+from tensorflow.keras.layers import Layer
 
 
 class DecodeDetections(Layer):
@@ -208,7 +208,7 @@ class DecodeDetections(Layer):
                 confidences = tf.expand_dims(batch_item[..., index], axis=-1)
                 class_id = tf.fill(
                     dims=tf.shape(confidences),
-                    value=tf.to_float(index))
+                    value=tf.cast(index, tf.float32))
                 box_coordinates = batch_item[..., -4:]
 
                 single_class = tf.concat(
